requirements.txt
transformers==4.31.0
torch==2.0.1
timm==0.9.2
Pillow==10.0.0
opencv-python==4.8.0.74
fastapi==0.100.1
uvicorn==0.23.2
gunicorn==21.2.0
python-multipart==0.0.6
cachetools==4.2.2